backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
requests==2.32.3

# ML/Search
sentence-transformers==3.0.1
torch==2.3.1
pinecone-client==5.0.1
Pillow==10.4.0
numpy==1.26.4
